Move git index search to internal
